use AndroidManifest to set the apiKey of gloop, splashscreen now stays only for the amount of time gloop needs to initialize, bug fixes

+++ app/src/main/java/io/gloop/drawed/BoardDetailFragment.java
@@ -29,7 +29,7 @@ public class BoardDetailFragment extends Fragment {
 
     private DrawingView drawView;
     private String currentColor = "#FF000000";
-    private float smallBrush, mediumBrush, largeBrush;
+    private int smallBrush, mediumBrush, largeBrush;
     private ImageView changeColorButton;
 
     private Board board;
@@ -95,7 +95,6 @@ public class BoardDetailFragment extends Fragment {
     public void onPause() {
         worker.stopWorker();
         super.onPause();
-
     }
 
     // opens a dialog on long press on the list item
